Add len method to video and hdf5 interfaces, make dunder test module to test dunder methods across all interfaces
